Moves cli.go
