+++ test/e2e/atlas_e2e_test_generator_test.go
@@ -115,7 +115,7 @@ func deleteProjectWithRetry(t *testing.T, projectID string) {
 
 func deleteKeys(t *testing.T, toDelete map[string]struct{}) {
 	t.Helper()
-	cliPath, err := test.AtlasCLIBin()
+	cliPath, err := AtlasCLIBin()
 
 	cmd := exec.Command(cliPath,
 		orgEntity,
