+++ test/e2e/helper_test.go
@@ -45,7 +45,7 @@ const (
 )
 
 func deployClusterForProject(projectID, tier, mDBVersion string, enableBackup bool) (string, string, error) {
-	cliPath, err := test.AtlasCLIBin()
+	cliPath, err := AtlasCLIBin()
 	if err != nil {
 		return "", "", err
 	}
@@ -104,7 +104,7 @@ func e2eTier() string {
 }
 
 func internalDeleteClusterForProject(projectID, clusterName string) error {
-	cliPath, err := test.AtlasCLIBin()
+	cliPath, err := AtlasCLIBin()
 	if err != nil {
 		return err
 	}
@@ -130,7 +130,7 @@ func internalDeleteClusterForProject(projectID, clusterName string) error {
 }
 
 func watchCluster(projectID, clusterName string) error {
-	cliPath, err := test.AtlasCLIBin()
+	cliPath, err := AtlasCLIBin()
 	if err != nil {
 		return err
 	}
@@ -151,7 +151,7 @@ func watchCluster(projectID, clusterName string) error {
 }
 
 func removeTerminationProtectionFromCluster(projectID, clusterName string) error {
-	cliPath, err := test.AtlasCLIBin()
+	cliPath, err := AtlasCLIBin()
 	if err != nil {
 		return err
 	}
@@ -190,7 +190,7 @@ func deleteClusterForProject(projectID, clusterName string) error {
 var errNoRegions = errors.New("no regions available")
 
 func newAvailableRegion(projectID, tier, provider string) (string, error) {
-	cliPath, err := test.AtlasCLIBin()
+	cliPath, err := AtlasCLIBin()
 	if err != nil {
 		return "", err
 	}
@@ -227,7 +227,7 @@ func newAvailableRegion(projectID, tier, provider string) (string, error) {
 }
 
 func RandClusterName() (string, error) {
-	n, err := test.RandInt(1000)
+	n, err := RandInt(1000)
 	if err != nil {
 		return "", err
 	}
@@ -238,7 +238,7 @@ func RandClusterName() (string, error) {
 }
 
 func RandProjectName() (string, error) {
-	n, err := test.RandInt(1000)
+	n, err := RandInt(1000)
 	if err != nil {
 		return "", err
 	}
@@ -281,7 +281,7 @@ func IsGov() bool {
 }
 
 func createProject(projectName string) (string, error) {
-	cliPath, err := test.AtlasCLIBin()
+	cliPath, err := AtlasCLIBin()
 	if err != nil {
 		return "", fmt.Errorf("%w: invalid bin", err)
 	}
@@ -310,7 +310,7 @@ func createProject(projectName string) (string, error) {
 }
 
 func deleteProject(projectID string) error {
-	cliPath, err := test.AtlasCLIBin()
+	cliPath, err := AtlasCLIBin()
 	if err != nil {
 		return err
 	}
