+++ test/e2e/kubernetes_operator_install_test.go
@@ -24,7 +24,6 @@ import (
 	"time"
 
 	"github.com/mongodb/atlas-cli-plugin-kubernetes/internal/kubernetes/operator/features"
-	"github.com/mongodb/atlas-cli-plugin-kubernetes/test"
 	akov2 "github.com/mongodb/mongodb-atlas-kubernetes/v2/pkg/api/v1"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -43,7 +42,7 @@ const (
 func TestKubernetesOperatorInstall(t *testing.T) {
 	req := require.New(t)
 
-	cliPath, err := test.PluginBin()
+	cliPath, err := PluginBin()
 	fmt.Println(cliPath)
 	req.NoError(err)
 	const contextPrefix = "kind-"
