+++ test/e2e/operator_helper_test.go
@@ -193,7 +193,7 @@ func (oh *operatorHelper) getOperatorSecretes(namespace string) ([]corev1.Secret
 func (oh *operatorHelper) installOperator(namespace, version string) error {
 	oh.t.Helper()
 
-	cliPath, err := test.PluginBin()
+	cliPath, err := PluginBin()
 	if err != nil {
 		return fmt.Errorf("unable to get atlasCli binary path: %w", err)
 	}
